pyproject.toml
[project]
name = "resume-parser-euriai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.80.0",
    "crewai-tools>=0.12.0",
    "pydantic>=2.8.2",
    "pypdf>=4.2.0",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "streamlit>=1.36.0",
]
